requirements.txt
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
fastapi==0.115.5
python-dotenv==1.0.1
joblib==1.4.2
pyparsing==3.2.0
pyarrow==17.0.0
pydantic==2.9.2
uvicorn== 0.30.6